Use v25 ag-grid npm packages
This is necessary because the library depends on a version of Feliz that
depends on React 17, but ag-grid-react@24 and ag-grid-community@24
depend on React 16. This means that the declared npm package
dependencies and the corresponding Nuget package dependencies are
incompatible.

There are some breaking changes to the ag-grid API (see
https://ag-grid.com/changelog/?fixVersion=25.0.0), but they do not
affect functions that were exposed via this library so there is no
breaking change to this library.

The change to this library is in
./fable/Feliz.AgGrid/Feliz.AgGrid.fsproj

This commit also makes various upgrades that I required to get this
working:
- .NET 5 -> 6
- Webpack 4 -> 5